Greatly reduce size on stack of addr tables
Limited systems, or otherwise conserative operating systems, may limit
the stack size of the thread that the node operates on. Most light
clients will not store more than a few hundred peers and even fewer will
have tried more than a handful during any given sync session. This
reduces the stack size of the table to a far more conservative size so
as to avoid a stack overflow.