Remove `CombindedAddr`
Revisiting the codebase after a while this struct is adding cluter
compared to the value it adds. No peers should be sending a standard
`Addr` message at this point, and if they are it should probably just be
ignored (maybe even disconnected in the future?). I can't think of any
reason this message would be sent except to fingerprint the recipient
by evaluating their behavior.