Remove `channel_messages`
Moves `ReaderMessage` into the `reader` module, as the only consumer is
`peer` and the only constructor of `ReaderMessage` is `reader`.

Removes the `GetHeaderConfig`, which appears to be some pointless
indirection like `CombinedAddr`. Case and point, the headers message
includes the protocol version, which should be made explicit.

The `MainThreadMessage` and `PeerThreadMessage` are just wrappers for
network communication in/out-bound, so these are moved into `network`.

This allows the opportunity to clean up some visibility modifiers as
well.